requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1

# Browser automation for Atria scraping
playwright==1.40.0

# Transcription
assemblyai==0.23.0

# AI/LLM APIs
anthropic>=0.75.0

# Google integrations
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
gspread==5.12.0

# Data processing
pandas==2.1.4
openpyxl==3.1.2

# Web server for webhooks
flask==3.0.0
gunicorn==21.2.0

# Utilities
loguru==0.7.2
python-dateutil==2.8.2
tenacity==8.2.3

# Async support
asyncio-throttle==1.0.2

# File handling
aiofiles==23.2.1

